@@ -14,7 +14,9 @@
         }
     ],
     "require": {
-        "php": ">=5.3.0"
+        "php": ">=8.1",
+        "ext-fileinfo": "*",
+      "ext-mbstring": "*"
     },
     "version": "2.9.8",
     "autoload": {
