@@ -22,9 +22,10 @@ function dump()
 
     // if last argument is boolean TRUE
     if ($arguments[count($arguments) - 1] === true)
-
+    {
         // set a flag telling the script that it needs to die() after output
         $stop = array_pop($arguments);
+    }
 
     // iterate through the arguments
     foreach ($arguments as $argument) {
@@ -37,14 +38,16 @@ function dump()
     }
 
     // if script needs to die() after output, die
-    if (isset($stop)) die();
+    if (isset($stop)) {
+        die();
+    }
 
 }
 ?>
 
 <!doctype html>
 
-<html>
+<html lang="en">
 
     <head>
 
@@ -76,15 +79,15 @@ function dump()
 
                     <?php
 
-                        function show_results() {
-
+                        function show_results(): void
+                        {
                             echo '<table class="results"><thead><tr><td colspan="2">Submitted values</td></tr></thead>';
 
                             foreach ($_POST as $key => $value) {
 
-                                if (strpos($key, 'name_') !== 0 && strpos($key, 'timer_') !== 0 && strpos($key, 'response_') !== 0)
-
+                                if (strpos($key, 'name_') !== 0 && strpos($key, 'timer_') !== 0 && strpos($key, 'response_') !== 0) {
                                     echo '<tr><th>' . $key . '</th><td>' . (is_array($value) ? '<pre>' . print_r($value, true) . '</pre>' : $value) . '</td></tr>';
+                                }
 
                             }
 
@@ -174,7 +177,7 @@ function dump()
 
                 <td valign="top">
 
-                    <?php if ($example != ''):?>
+                    <?php if (!empty($example)):?>
 
                     <ul class="tabs float clearfix">
 
@@ -190,13 +193,13 @@ function dump()
 
                     <?php for ($i = 0; $i < 5; $i++):?>
 
-                    <div class="tab clearfix"<?php echo $i == 0 ? ' style="display:block"' : ''?>>
+                    <div class="tab clearfix"<?php echo $i === 0 ? ' style="display:block"' : ''?>>
 
-                        <?php if ($i == 0):?>
+                        <?php if ($i === 0):?>
 
                             <?php require 'includes/' . $current_example . '-' . $current_template . '.php'?>
 
-                        <?php elseif ($i == 1):?>
+                        <?php elseif ($i === 1):?>
 
                         <?php
 
@@ -224,9 +227,9 @@ function dump()
 
                         ?></code></pre>
 
-                        <?php elseif ($i == 2):?>
+                        <?php elseif ($i === 2):?>
 
-                            <?php if (strpos($current_template, 'custom') === false):?>
+                            <?php if (!str_contains($current_template, 'custom')):?>
 
                                 <p>In this example, the output is automatically generated by Zebra_Form's <strong>render</strong> method.</p>
 
@@ -244,7 +247,7 @@ function dump()
 
                             <?php endif?>
 
-                        <?php elseif ($i == 3):?>
+                        <?php elseif ($i === 3):?>
 
                         <?php
 
@@ -260,7 +263,7 @@ function dump()
 
                         <?php endif?>
 
-                        <?php if ($i == 0):?>
+                        <?php if ($i === 0):?>
 
                         <ul class="notes default">
                         	<li>
