@@ -31,7 +31,7 @@ class Zebra_Form_Password extends Zebra_Form_Control
      *  $form->render();
      *  </code>
      *
-     *  @param  string  $id             Unique name to identify the control in the form.
+     *  @param string $id             Unique name to identify the control in the form.
      *
      *                                  The control's <b>name</b> attribute will be the same as the <b>id</b> attribute!
      *
@@ -47,9 +47,9 @@ class Zebra_Form_Password extends Zebra_Form_Control
      *                                  echo $my_password;
      *                                  </code>
      *
-     *  @param  string  $default        (Optional) Default value of the password field.
+     *  @param string $default        (Optional) Default value of the password field.
      *
-     *  @param  array   $attributes     (Optional) An array of attributes valid for
+     *  @param array $attributes     (Optional) An array of attributes valid for
      *                                  {@link http://www.w3.org/TR/REC-html40/interact/forms.html#h-17.4 input}
      *                                  controls (size, readonly, style, etc)
      *
@@ -111,7 +111,7 @@ class Zebra_Form_Password extends Zebra_Form_Control
      *
      *  @return void
      */
-    function __construct($id, $default = '', $attributes = '')
+    public function __construct(string $id, string $default = '', array $attributes = [])
     {
     
         // call the constructor of the parent class
@@ -121,11 +121,9 @@ class Zebra_Form_Password extends Zebra_Form_Control
         // these attributes are private for this control and are for internal use only
         // and will not be rendered by the _render_attributes() method
         $this->private_attributes = array(
-
             'default_value',
             'disable_xss_filters',
             'locked',
-
         );
 
         // set the default attributes for the button control
@@ -166,11 +164,9 @@ class Zebra_Form_Password extends Zebra_Form_Control
      *
      *  @return string  The control's HTML code
      */
-    function toHTML()
+    public function toHTML(): string
     {
-
-        return '<input ' . $this->_render_attributes() . ($this->form_properties['doctype'] == 'xhtml' ? '/' : '') . '>';
-
+        return '<input ' . $this->_render_attributes() . ($this->form_properties['doctype'] === 'xhtml' ? '/' : '') . '>';
     }
 
 }
