@@ -105,7 +105,7 @@ class Zebra_Form_Radio extends Zebra_Form_Control
      *  if you plan on adding/removing values dynamically, from JavaScript, you will have to call the
      *  {@link Zebra_Form_Control::disable_spam_filter() disable_spam_filter()} method to prevent that from happening!</samp>
      *
-     *  @param  string  $id             Unique name to identify the control in the form.
+     *  @param string $id             Unique name to identify the control in the form.
      *
      *                                  The control's <b>name</b> attribute will be as indicated by <i>$id</i>
      *                                  argument while the control's <b>id</b> attribute will be <i>$id</i> followd by an
@@ -134,7 +134,7 @@ class Zebra_Form_Radio extends Zebra_Form_Control
      *
      *  @param  mixed   $value          Value of the radio button.
      *
-     *  @param  array   $attributes     (Optional) An array of attributes valid for
+     *  @param array $attributes     (Optional) An array of attributes valid for
      *                                  {@link http://www.w3.org/TR/REC-html40/interact/forms.html#h-17.4 input}
      *                                  controls (disabled, readonly, style, etc)
      *
@@ -161,7 +161,7 @@ class Zebra_Form_Radio extends Zebra_Form_Control
      *
      *  @return void
      */
-    function __construct($id, $value, $attributes = '')
+    public function __construct(string $id, mixed $value, array $attributes = [])
     {
 
         // call the constructor of the parent class
@@ -171,11 +171,9 @@ class Zebra_Form_Radio extends Zebra_Form_Control
         // these attributes are private for this control and are for internal use only
         // and will not be rendered by the _render_attributes() method
         $this->private_attributes = array(
-
             'disable_spam_filter',
             'disable_xss_filters',
             'locked',
-
         );
 
         // set the default attributes for the radio button control
@@ -217,11 +215,9 @@ class Zebra_Form_Radio extends Zebra_Form_Control
      *
      *  @return string  The control's HTML code
      */
-    function toHTML()
+    public function toHTML(): string
     {
-
-        return '<input ' . $this->_render_attributes() . ($this->form_properties['doctype'] == 'xhtml' ? '/' : '') . '>';
-
+        return '<input ' . $this->_render_attributes() . ($this->form_properties['doctype'] === 'xhtml' ? '/' : '') . '>';
     }
 
 }
