@@ -31,7 +31,7 @@ class Zebra_Form_Textarea extends Zebra_Form_Control
      *  $form->render();
      *  </code>
      *
-     *  @param  string  $id             Unique name to identify the control in the form.
+     *  @param string $id             Unique name to identify the control in the form.
      *
      *                                  The control's <b>name</b> attribute will be the same as the <b>id</b> attribute!
      *
@@ -47,9 +47,9 @@ class Zebra_Form_Textarea extends Zebra_Form_Control
      *                                  echo $my_textarea;
      *                                  </code>
      *
-     *  @param  string  $default        (Optional) Default value of the textarea.
+     *  @param string $default        (Optional) Default value of the textarea.
      *
-     *  @param  array   $attributes     (Optional) An array of attributes valid for
+     *  @param array $attributes     (Optional) An array of attributes valid for
      *                                  <b>{@link http://www.w3.org/TR/REC-html40/interact/forms.html#h-17.7 textarea}</b>
      *                                  controls (rows, cols, style, etc)
      *
@@ -76,7 +76,7 @@ class Zebra_Form_Textarea extends Zebra_Form_Control
      *
      *  @return void
      */
-    function __construct($id, $default = '', $attributes = '')
+    public function __construct(string $id, string $default = '', array $attributes = [])
     {
     
         // call the constructor of the parent class
@@ -136,7 +136,7 @@ class Zebra_Form_Textarea extends Zebra_Form_Control
      *
      *  @return string  The control's HTML code
      */
-    function toHTML()
+    public function toHTML(): string
     {
 
         // get private attributes
